requirements.txt
weaviate-client
tqdm
pandas
numpy
flask